refactor(dashboard): extract Mastio vault + oidc sub-routers (F-B-201 PR-9/10)
Move the four /proxy/vault routes (page + save + test + migrate-keys)
plus the two /proxy/oidc routes (start + callback) out of the ~3500-LOC
mcp_proxy/dashboard/router.py god-object into two dedicated modules,
each mounted via include_router.

Same pattern as PR-2..PR-8. CSRF gate, approval-hook intercept on
migrate-keys and the SSRF guard on /vault/test all preserved verbatim.
OIDC handshake state machine unchanged.

Pruned now-unused router.py imports: ACTION_VAULT_MIGRATE_KEYS,
_load_display_name, set_session. tests/test_proxy_dashboard_oidc.py
updated to import _oidc_redirect_uri from the new sub-router.

F-B-201 progress: 9/10 PR.
